requirements.txt
djangorestframework
django-filter
djangorestframework-simplejwt
drf-yasg
PyYAML